+++ lib/defs/api.ts
@@ -2,6 +2,9 @@
  * Welcome to ng2tree
  */
 
+export interface IAllowDropFn {
+  (element:any, to:{parent:ITreeNode, index:number}):boolean;
+}
 /**
 * This is the interface of the options input of the tree.
 * See docs for more detailed explanations
@@ -36,9 +39,17 @@ export interface ITreeOptions {
     */
    actionMapping?: any;
    /**
-    * Allow drag and drop on the tree. Default: false
+    * Allow dragging tree nodes. Default: false
     */
    allowDrag?: boolean;
+   /**
+    * Allow drop on the tree.
+    * Either boolean value, or a function that takes the dragged element and drop location (parent, index)
+    * and returns a boolean value.
+    * Will be called when dragging over the nodes or between nodes, and allow to prevent the drop style
+    * Default: true
+    */
+   allowDrop?: boolean | IAllowDropFn;
    /**
    * Specify padding per node instead of children padding (to allow full row select for example)
    */